# openapi.json
{
...
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/v1/captchas/@createCaptchaBySms": {
      "post": {
        "operationId": "createCaptchaBySms",
        "summary": "",
        "description": "Create captcha by sms",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaptchaBySmsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The captcha by sms has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptchaBySmsResult"
                }
              }
            }
          }
        },
        "tags": ["captcha"]
      }
    },
    "/auth/v1/captchas/@createCaptchaByEmail": {
      "post": {
        "operationId": "createCaptchaByEmail",
        "summary": "",
        "description": "Create captcha by email",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaptchaByEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The captcha by email has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptchaByEmailResult"
                }
              }
            }
          }
        },
        "tags": ["captcha"]
      }
    },
    "/auth/v1/captchas/@createCaptchaByPhoto": {
      "post": {
        "operationId": "createCaptchaByPhoto",
        "summary": "",
        "description": "Create captcha by photo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaptchaByPhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The captcha by photo has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptchaByPhotoResult"
                }
              }
            }
          }
        },
        "tags": ["captcha"]
      }
    },
...
  },
  "info": {
    "title": "Auth API Server",
    "description": "Auth API for auth service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CaptchaPurpose": {
        "type": "string",
        "description": "用途",
        "enum": ["REGISTER", "LOGIN", "RESET_PASSWORD", "UPDATE_PHONE", "UPDATE_EMAIL"]
      },
      "CreateCaptchaBySmsDto": {
        "type": "object",
        "properties": {
          "purpose": {
            "$ref": "#/components/schemas/CaptchaPurpose"
          },
          "phone": {
            "type": "string",
            "description": "手机号，必填"
          },
          "dialingPrefix": {
            "type": "string",
            "description": "区号，必填"
          },
          "scope": {
            "type": "string",
            "description": "scope"
          }
        },
        "required": ["purpose", "phone", "dialingPrefix", "scope"]
      },
      "CaptchaBySmsResult": {
        "type": "object",
        "properties": {
          "purpose": {
            "$ref": "#/components/schemas/CaptchaPurpose"
          },
          "phone": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "description": "scope"
          },
          "dialingPrefix": {
            "type": "string",
            "description": "区号"
          },
          "expireAt": {
            "format": "date-time",
            "type": "string",
            "description": "过期时间"
          }
        },
        "required": ["purpose", "phone", "scope", "expireAt"]
      },
      "CreateCaptchaByEmailDto": {
        "type": "object",
        "properties": {
          "purpose": {
            "$ref": "#/components/schemas/CaptchaPurpose"
          },
          "email": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "description": "scope"
          }
        },
        "required": ["purpose", "email", "scope"]
      },
      "CaptchaByEmailResult": {
        "type": "object",
        "properties": {
          "purpose": {
            "$ref": "#/components/schemas/CaptchaPurpose"
          },
          "email": {
            "type": "string"
          },
          "scope": {
            "type": "string",
            "description": "scope"
          },
          "expireAt": {
            "format": "date-time",
            "type": "string",
            "description": "过期时间"
          }
        },
        "required": ["purpose", "email", "scope", "expireAt"]
      },
      "CreateCaptchaByPhotoDto": {
        "type": "object",
        "properties": {
          "purpose": {
            "$ref": "#/components/schemas/CaptchaPurpose"
          },
          "key": {
            "type": "string",
            "description": "key"
          },
          "scope": {
            "type": "string",
            "description": "scope"
          }
        },
        "required": ["purpose", "key", "scope"]
      },
      "CaptchaByPhotoResult": {
        "type": "object",
        "properties": {
          "purpose": {
            "$ref": "#/components/schemas/CaptchaPurpose"
          },
          "capchaGifHex": {
            "type": "string",
            "description": "gif 图片内容"
          },
          "scope": {
            "type": "string",
            "description": "scope"
          },
          "key": {
            "type": "string",
            "description": "key"
          },
          "expireAt": {
            "format": "date-time",
            "type": "string",
            "description": "过期时间"
          }
        },
        "required": ["purpose", "capchaGifHex", "scope", "key", "expireAt"]
      },
...
    }
  }
}